// LinkERP/ClientApp/src/app/models/pos/lbs-sop-mixture.ts
import { BaseEntity } from "../base-entity";
export class LbsSopMixture extends BaseEntity {
    MixtureID: any;
    CompanyID: any;
    PrescriptionID: any;
    PrescriptionLineNumber: any;
    ProductID: any;
    ProductCode: any
    ProductName: any;
    Quantity: any;
    UnitPrice:any;
    UnitCost: any;
    Total: any;
    Tax: any;
    Processed: any;
    WareHouseID: any;
}